[test] Add JUnit test for UserRepository
- Add JUnit test code for UserRepository
- Add UserRepository query method(name, email)
