+++ src/main/java/gg/boardgame/bdgg/db/UserRepository.java
@@ -2,6 +2,12 @@ package gg.boardgame.bdgg.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findById(Long id);
+    Optional<User> findByName(String name);
+    Optional<User> findByEmail(String email);
 
 }
